Fix warnings.
I know it's a minor thing, but it's kind of nice to have things compile without
warnings. This should make it easier to find other issues that show up. That
is, if you see a warning now, it's likely that it is new code that added it.

This corresponds to patch 3196483 (with two additions).