Fix interaction with vendors in the iOS version.
This allows for a better experience when dealing with vendors.
For buying: Create buttons that actually point to the items name, instead of "Choice X".
For selling: Actually get the list of items in our inventory instead of just listing everything.

The only downside is that for the Guild shop does show sextant. This is
unfortunate, but at the same time, I can live with that.