Fix compilation to work on iOS again.
I understand the optimizations that were done a little while ago seemed like a
good idea, but they broke the iOS version. Sometimes the abstraction is a good
thing. Also, the places that I reverted them back (initializing and turning on
and off music) are not bottlenecks in the code.