Drive all display updates from a very limited number of places.
 - Add screenSwapBuffers (ala glxSwapBuffers).
 - Remove screenRedrawScreen, screenRedrawTextArea, & timerQueueEmpty.

For Allegro the primary updates are in EventHandler::run() & sleep().
SDL still uses its screenRefreshThread.

ProgressBar does its updates manually since it exists before the
EventHandler is running.  This may not be needed with SDL now, but in the
future that implementation may also use a timer event rather than a thread.