Provide an alternative shadowcasting line of sight algorithm.
Overall the performance is slightly worse, being around 3x slower in open
areas, but 3x faster in tightly enclosed ones.

The primary advantages are that the viewport size is not hardcoded and tile
distances from the viewer are available.  As these features are not
currently needed this algorithm is disabled.

The original code is from https://fadden.com/tech/ShadowCast.cs.txt.