game.cpp: Restart the turn Pass timer when moving to a new map.
This fixes a camping bug where a forced Pass would occur after resting.
A CombatController::finishTurn() here acts as a combat Victory! and an extra
exitToParentMap() is done.

The root cause is that key commands do not call gameStampCommandTime()
until the end of switch(key), where the command is known to be valid.
A proper fix would be to run the resting scene outside of the key handler
function.